+++ NEWS.md
@@ -1,7 +1,7 @@
 # dbparser 1.0.2
 * Fix zip file location issue
 * Replace Secondary and third keys columns from drug framework with *other_keys* column that contains any other keys that might exist in addition to the primary key
-* Add **average-mass**, **monoisotopic-mass** attributes.
+* Add **average-mass**, **monoisotopic-mass** and **calculated-properties** attributes.
 
 
 # dbparser 1.0.1